requirements.txt
langchain>=0.0.148
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
pandas>=1.3.0
tqdm>=4.64.0
